+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Version 1.5.2 [2022-03-17]
 APP: **HAT_RunAnalyzer_HMC.py**
     - FIX: bugs related to the operational mode.
 APP: **HAT_RunPublisher_HMC_Maps.py**
+	- ADD: grid static information from file or using the basic geographical information (xll_corner, yll_corner, rows, cols and cell_size).
     - FIX: bugs related to the operational mode.
     - FIX: bugs related to the mandatory/ancillary static data (terrain, flow_directions and river_network).
 APP: **HAT_RunPublisher_HMC_TimeSeries.py**